==> code/01_Yelp_Restaurant_Data_Gathering.ipynb <==
# Data Gathering
## Yelp API Test Run

# Imports
import requests
import pandas as pd
import time

# Defining our key
yelp_api_key = '7eY2PBhjxgtq0f7519bHqvwWeIRxap1qWw7UBUb9gB4eAPOtOQq9szQPmELDWVx-w7KIaeOCJN4R8Yrs25kt2-pQSJQIV03QmZ7NB-0e14_kxxEutPSenV4OK_1FXnYx'

# Define the endpoint
yelp_url = 'https://api.yelp.com/v3/businesses/search'

# Define the header
header = {'Authorization': f'bearer {yelp_api_key}'}

# Defining the parameters of the search
params = {'term': 'restaurants',
          'limit': 20,
          'location': '06905'
}

# Making the request
response = requests.get(url = yelp_url, params = params, headers = header)

# Checking response status
response.status_code

# Converting the JSON to a dictionary
r_dict = response.json()

r_dict = r_dict['businesses']

new_dict = {}

# Parsing the data
for i in range(len(r_dict)):
    try: 
        new_dict[i] = {
            'id': r_dict[i]['id'],
            'name': r_dict[i]['name'],
            'review_count': r_dict[i]['review_count'],
            'rating': r_dict[i]['rating'],
            'city': r_dict[i]['location']['city'],
            'zip_code': r_dict[i]['location']['zip_code'],
            'state': r_dict[i]['location']['state'],   
            'price': r_dict[i]['price']
        } 
    except: continue

df = pd.DataFrame(new_dict)

df = df.T

# Making sure we have what we want
df.head()

df.shape

## Yelp API Function

# Creating a list of zip codes
zip_codes=[
"06230", 
"06231",
"06232",
"06401",
"06278",
"06001",
"06330",
"06750",
"06063",
"06403",
"06037",
"06524",
"06801",
"06751",
"06002",
"06043",
"06404",
"06334",
"06405",
"06601",
"06602",
"06604",
"06605",
"06606",
"06607",
"06608",
"06610",
"06752",
"06010",
"06011",
"06016",
"06804",
"06234",
"06013",
"06018",
"06331",
"06019",
"06020",
"06409",
"06332",
"06235",
"06408",
"06411",
"06412",
"06413",
"06414",
"06415",
"06021",
"06022",
"06237",
"06753",
"06754",
"06807",
"06238",
"06416",
"06810",
"06811",
"06813",
"06816",
"06817",
"06239",
"06820",
"06241",
"06417",
"06418",
"06422",
"06023",
"06024",
"06025",
"06026",
"06423",
"06424",
"06108",
"06118",
"06128",
"06138",
"06027",
"06243",
"06333",
"06088",
"06028",
"06244",
"06242",
"06612",
"06029",
"06082",
"06083",
"06426",
"06824",
"06825",
"06828",
"06031",
"06030",
"06032",
"06034",
"06335",
"06755",
"06829",
"06336",
"06033",
"06756",
"06035",
"06838",
"06830",
"06831",
"06836",
"06246",
"06340",
"06349",
"06437",
"06438",
"06439",
"06514",
"06517",
"06518",
"06247",
"06350",
"06101",
"06102",
"06103",
"06104",
"06105",
"06106",
"06112",
"06114",
"06115",
"06120",
"06126",
"06132",
"06134",
"06141",
"06142",
"06143",
"06144",
"06145",
"06146",
"06147",
"06151",
"06152",
"06153",
"06154",
"06155",
"06156",
"06160",
"06183",
"06791",
"06440",
"06248",
"06441",
"06442",
"06351",
"06757",
"06419",
"06758",
"06039",
"06249",
"06339",
"06759",
"06443",
"06040",
"06041",
"06042",
"06045",
"06250",
"06251",
"06444",
"06447",
"06338",
"06450",
"06451",
"06456",
"06762",
"06455",
"06457",
"06459",
"06460",
"06461",
"06467",
"06468",
"06353",
"06469",
"06354",
"06763",
"06355",
"06770",
"06050",
"06051",
"06052",
"06053",
"06840",
"06812",
"06057",
"06501",
"06502",
"06503",
"06504",
"06505",
"06506",
"06507",
"06508",
"06509",
"06510",
"06511",
"06512",
"06513",
"06515",
"06519",
"06520",
"06525",
"06530",
"06531",
"06533",
"06534",
"06536",
"06320",
"06776",
"06777",
"06111",
"06131",
"06470",
"06357",
"06058",
"06471",
"06059",
"06254",
"06060",
"06255",
"06473",
"06359",
"06256",
"06778",
"06472",
"06850",
"06851",
"06852",
"06853",
"06854",
"06855",
"06856",
"06857",
"06859",
"06860",
"06360",
"06370",
"06779",
"06870",
"06371",
"06372",
"06475",
"06373",
"06477",
"06478",
"06379",
"06781",
"06061",
"06374",
"06062",
"06479",
"06782",
"06258",
"06259",
"06064",
"06480",
"06365",
"06712",
"06260",
"06375",
"06262",
"06896",
"06875",
"06876",
"06877",
"06878",
"06065",
"06481",
"06067",
"06263",
"06783",
"06420",
"06068",
"06482",
"06264",
"06483",
"06069",
"06484",
"06784",
"06070",
"06071",
"06072",
"06487",
"06073",
"06785",
"06376",
"06266",
"06074",
"06267",
"06488",
"06489",
"06890",
"06075",
"06076",
"06077",
"06901",
"06902",
"06903",
"06904",
"06905",
"06906",
"06907",
"06911",
"06912",
"06920",
"06921",
"06926",
"06927",
"06377",
"06491",
"06378",
"06268",
"06269",
"06614",
"06615",
"06078",
"06080",
"06079",
"06380",
"06081",
"06786",
"06787",
"06277",
"06084",
"06790",
"06792",
"06611",
"06382",
"06085",
"06066",
"06383",
"06384",
"06492",
"06493",
"06793",
"06794",
"06701",
"06702",
"06703",
"06704",
"06705",
"06706",
"06708",
"06710",
"06720",
"06721",
"06722",
"06723",
"06724",
"06749",
"06385",
"06795",
"06387",
"06089",
"06796",
"06090",
"06107",
"06110",
"06117",
"06119",
"06127",
"06133",
"06137",
"06091",
"06516",
"06388",
"06092",
"06093",
"06498",
"06883",
"06880",
"06881",
"06889",
"06109",
"06129",
"06226",
"06279",
"06897",
"06094",
"06280",
"06006",
"06095",
"06096",
"06098",
"06716",
"06798",
"06281",
"06282",
"06389",]

# This function will do the process above for each zip code in a list of zip codes we input
def yelp_request(key, zip_codes, term = 'restaurants', limit =50):
    
    # Defining the endpoint and header
    yelp_url = 'https://api.yelp.com/v3/businesses/search'
    header = header = {'Authorization': f'bearer {key}'}
    
    # Defining an empty dictionary to store our restaurant data in
    r_dict = {}
    
    # Defining an empty data frame to append to
    df = pd.DataFrame(columns = ['id', 'name', 'review_count', 'rating', 'city', 'zip_code', 'state', 'price'])
    
    # Looping through each zip code to run a request for each
    for i in range(len(zip_codes)):
        
        # Defining our parameters
        params = {'term': 'restaurants',
          'limit': limit,
          'location': zip_codes[i]
        }
        
        # Making the request
        response = requests.get(url = yelp_url, params = params, headers = header)
        
        # Checking the status code
        #if response.status_code != 200:
            #raise
        
        # Converting JSON to a dictionary
        temp_dict = response.json()
        temp_dict = temp_dict['businesses']
        
        # Parsing the data 
        new_dict = {}
        for i in range(len(temp_dict)):
            try: 
                new_dict[i] = {
                    'id': temp_dict[i]['id'],
                    'name': temp_dict[i]['name'],
                    'review_count': temp_dict[i]['review_count'],
                    'rating': temp_dict[i]['rating'],
                    'city': temp_dict[i]['location']['city'],
                    'zip_code': temp_dict[i]['location']['zip_code'],
                    'state': temp_dict[i]['location']['state'],   
                    'price': temp_dict[i]['price']
            } 
            except: continue

        # Creating a data frame
        temp_df = pd.DataFrame(new_dict)
        temp_df = temp_df.T

        # Adding to complete data frame
        df = df.append(temp_df)
        
        # Waiting a second before scraping again
        time.sleep(1)
    
    # Saving to a csv
    df.to_csv('../data/yelp_2.csv', index=False)
    
    # Returning the final data frame
    return df

# We ran this four times due to our limit of 5000 documents per day
df = yelp_request(yelp_api_key, zip_codes[200:300])

# Checking the size
df.shape

# Dropping duplicates
df = df.drop_duplicates()

df.shape

# Taking a look at the data
df.head()
