==> code/04_CT_Income_Data_Cleaning.ipynb <==
# Data Cleaning: CT Income

# Imports
import pandas as pd 

# Reading in the CT income data
df = pd.read_csv('../data/ct_income_raw.csv')

# Taking a look at the data
df.head()

# Checking the data types
df.dtypes

**Note:** the data types are strings when they should be integers. The commas are likely what is causing this.

# Getting rid of commas and casting the values to integers
df['median_household_income'] = df['median_household_income'].str.replace(',', '').astype(int)
df['mean_household_income'] = df['mean_household_income'].str.replace(',', '').astype(int)
df['per_capita_income'] = df['per_capita_income'].str.replace(',', '').astype(int)

## Median Household Income Buckets

# 1: $1 under $25,000
# 2: $25,000 under $50,000
# 3: $50,000 under $75,000
# 4: $75,000 under $100,000
# 5: $100,000 under $200,000
# 6: $200,000 or more

mask0_25k = (df['median_household_income'] > 0) & (df['median_household_income'] <= 25_000)
mask25k_50k = (df['median_household_income'] >= 25_001) & (df['median_household_income'] <= 50_000)
mask50k_75k = (df['median_household_income'] >= 50_001) & (df['median_household_income'] <= 75_000)
mask75k_100k = (df['median_household_income'] >= 75_001) & (df['median_household_income'] <= 100_000)
mask100k_200k = (df['median_household_income'] >= 100_001) & (df['median_household_income'] <= 200_000)
mask200k_more = (df['median_household_income'] >= 200_001) 

bucket1 = df.loc[mask0_25k]
bucket2 = df.loc[mask25k_50k]
bucket3 = df.loc[mask50k_75k]
bucket4 = df.loc[mask75k_100k]
bucket5 = df.loc[mask100k_200k]
bucket6 = df.loc[mask200k_more]

bucket1['median_bucket'] = 1
bucket2['median_bucket'] = 2
bucket3['median_bucket'] = 3
bucket4['median_bucket'] = 4
bucket5['median_bucket'] = 5
bucket6['median_bucket'] = 6

df_bucket = pd.concat([bucket1,bucket2,bucket3,bucket4,bucket5,bucket6,], axis = 0)

df_bucket.shape

# Sort by city 
df_bucket.sort_values(by = 'city', inplace = True)

df_bucket.head()

## Mean Household Income Buckets

# 1: $1 under $25,000
# 2: $25,000 under $50,000
# 3: $50,000 under $75,000
# 4: $75,000 under $100,000
# 5: $100,000 under $200,000
# 6: $200,000 or more

mask0_25k = (df_bucket['mean_household_income'] > 0) & (df_bucket['mean_household_income'] <= 25_000)
mask25k_50k = (df_bucket['mean_household_income'] >= 25_001) & (df_bucket['mean_household_income'] <= 50_000)
mask50k_75k = (df_bucket['mean_household_income'] >= 50_001) & (df_bucket['mean_household_income'] <= 75_000)
mask75k_100k = (df_bucket['mean_household_income'] >= 75_001) & (df_bucket['mean_household_income'] <= 100_000)
mask100k_200k = (df_bucket['mean_household_income'] >= 100_001) & (df_bucket['mean_household_income'] <= 200_000)
mask200k_more = (df_bucket['mean_household_income'] >= 200_001) 

bucket1 = df_bucket.loc[mask0_25k]
bucket2 = df_bucket.loc[mask25k_50k]
bucket3 = df_bucket.loc[mask50k_75k]
bucket4 = df_bucket.loc[mask75k_100k]
bucket5 = df_bucket.loc[mask100k_200k]
bucket6 = df_bucket.loc[mask200k_more]

bucket1['mean_bucket'] = 1
bucket2['mean_bucket'] = 2
bucket3['mean_bucket'] = 3
bucket4['mean_bucket'] = 4
bucket5['mean_bucket'] = 5
bucket6['mean_bucket'] = 6

df_bucket = pd.concat([bucket1,bucket2,bucket3,bucket4,bucket5,bucket6,], axis = 0)

df_bucket.head()

df_bucket.sort_values(by = 'city', inplace = True)

df_bucket.head()

#### Export

df_bucket.to_csv('../data/ct_income_clean.csv', index=False)

## Per Capita Income Buckets

df_bucket['per_capita_income'].describe()
