==> code/05_Combining_Data.ipynb <==
# Combining Data

# Imports
import pandas as pd

# Reading in data
irs = pd.read_csv('../data/irs_clean_2.csv')
city = pd.read_csv('../data/avg_price_restaurants.csv')
city_2 = pd.read_csv('../data/perc_price_category.csv')

## IRS Data

### Grouping IRS Data by City

irs.head()

# Creating a map for the total of returns per city
total_returns_map = irs.groupby('city').sum()['n_returns']

# Creating a column from the map for the total number of returns per city
irs['total_n_returns'] = irs['city'].map(total_returns_map)

irs.head()

# Calculating the weighted average of the income bucket per city weighted by the number of returns
irs['y'] = irs['n_returns'] / irs['total_n_returns'] * irs['income_bucket']

# Grouping by city to merge with restaurant data
y = pd.DataFrame(irs.groupby('city').sum()['y'].round(0))

y.head()

# Resetting the index so that the index is not cities
y.reset_index(inplace=True)

### Merging IRS Data with Average Price Restaurant Data

city.head()

city.shape

y.shape

**Note:** the shapes do not match up, so some cities in our city data are not represented in the irs data.

# Checking which cities are missing from the IRS data
missing_city = []
for c in city['city'].values:
    if c not in y['city'].values:
        missing_city.append(c)

missing_city

# Checking which cities are missing from the city data
missing_city_2 = []
for c in y['city'].values:
    if c not in city['city'].values:
        missing_city_2.append(c)

missing_city_2

**Note:** some cities, such as E. Hartford (East Hartford) may be typed out differently. There are also some that have random caps ("South WIndsor"). Other cities may just actually not be contained in both datasets. I will lower case all and clean up some of the city names to make sure we line up the cities that are in both datasets.

# Capitalizing the first letter of every word, which is how cities are represented in the irs data
city['city'] = city['city'].str.title()

# Dropping cities in the restaurant data that are not in the irs data
city = city.loc[city['city'].isin(y['city'])]

# Dropping cities in the irs data that are not in the restaurant data
y = y.loc[y['city'].isin(city['city'])]

city.shape

y.shape

**Note:** the restaurant data still contains two additional cities. These are likely duplicates.

# Dropping duplicates
city = city.drop_duplicates(subset = 'city')

city.shape

# Merging both data frames on city
df_1 = city.merge(y, on='city')

# Taking a look at the data
df_1.head()

#### Export

df_1.to_csv('../data/combined_1.csv', index=False)

### Merging IRS Data with Price Category Restaurant Data

city_2.head()

city_2.shape

y.shape

**Note:** the shapes do not match up, so we will go through the same process we did for the other data.

# Capitalizing the first letter of every word, which is how cities are represented in the irs data
city_2['city'] = city_2['city'].str.title()

# Dropping cities in the restaurant data that are not in the irs data
city_2 = city_2.loc[city_2['city'].isin(y['city'])]

# Dropping cities in the irs data that are not in the restaurant data
y = y.loc[y['city'].isin(city_2['city'])]

# Dropping duplicates
city_2 = city_2.drop_duplicates(subset = 'city')

city_2.shape

y.shape

# Merging both data frames on city
df_2 = city_2.merge(y, on='city')

# Taking a look at the data
df_2.head()

#### Export

df_2.to_csv('../data/combined_2.csv', index=False)

## CT Income Data

# Reading in the data
ct_df = pd.read_csv('../data/ct_income_clean.csv')
city = pd.read_csv('../data/avg_price_restaurants.csv')
city_2 = pd.read_csv('../data/perc_price_category.csv')

# Taking a look at the data
ct_df.head()

### Merging CT Data with Average Price Restaurant Data

city.head()

city.shape

ct_df.shape

# Shapes do not match, so we will go through the same process as before
city = city.loc[city['city'].isin(ct_df['city'])]
ct_df = ct_df.loc[ct_df['city'].isin(city['city'])]

city.shape

ct_df.shape

# Merging the data frames on city
df_3 = city.merge(ct_df, on='city')

# Taking a look at the data
df_3.head()

#### Export

df_3.to_csv('../data/combined_3.csv', index=False)

### Merging CT Data with Price Category Restaurant Data

city_2.head()

city_2.shape

ct_df.shape

# Shapes do not match, so we will go through the same process as before
city_2 = city_2.loc[city_2['city'].isin(ct_df['city'])]
ct_df = ct_df.loc[ct_df['city'].isin(city_2['city'])]

city_2.shape

ct_df.shape

# Merging the data frames on city
df_4 = city_2.merge(ct_df, on='city')

# Taking a look at the data
df_4.head()

#### Export

df_4.to_csv('../data/combined_4.csv', index=False)
