==> code/07_Modeling.ipynb <==
# Modeling

# Imports
import pandas as pd
from sklearn.linear_model import LogisticRegression, LinearRegression
from sklearn.model_selection import train_test_split, GridSearchCV

## IRS Data

## Average Price Model

# Reading in the data
df_1 = pd.read_csv('../data/combined_1.csv')

df_1.head()

# Getting our baseline
df_1['y'].value_counts(normalize=True)

**Note:** This is not good. We have highly imbalanced classes, and one of the classes (1.0) is not represented at all. This is a result of the way we binned the income categories (as a weighted average of the price category weighted by the number of returns in each category). We will try to run a model, but it will likely predict all 3's, as this is the majority class.

# Defining our feature vector and prediction vector
X = df_1.drop(columns=['city', 'y'])
y = df_1['y']

# Train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, stratify=y)

# Instantiating a logistic regression model
lr = LogisticRegression(solver='liblinear')

# Defining parameters to search over
lr_params = {
    'C': [.01, .1, 1],
    'penalty': ['l1', 'l2', 'elasticnet', 'none']
}

# Instantiating a grid search 
gs = GridSearchCV(lr, lr_params, cv=5, scoring = 'accuracy')

# Fitting the model
gs.fit(X_train, y_train)

print(f'Best Parameters: {gs.best_params_}')

print(f'Best r-squared Score: {gs.best_score_}')

# Saving the best model
lr = gs.best_estimator_

# Checking the score on the test set
lr.score(X_test, y_test)

# Checking the predictions
lr.predict(X_test)

**Note:** as expected, our model predicted 3's for everything, so this is useless for predicting affluence. We will use the CT income data as our predictions instead of the IRS data.

## CT Income Data

## Average Price Models

# Reading in the data
df_1 = pd.read_csv('../data/combined_3.csv')

df_1.head()

### Linear Regression Model Using Per Capita Income

X = df_1[['price']]
y = df_1['per_capita_income']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

lr = LinearRegression()
lr.fit(X_train, y_train)

print(f'Score on the training set: {lr.score(X_train, y_train)}')
print(f'Score on the test set: {lr.score(X_test, y_test)}')

### Linear Regression Model Using Median Household Income

X = df_1[['price']]
y = df_1['median_household_income']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

lr = LinearRegression()
lr.fit(X_train, y_train)

print(f'Score on the training set: {lr.score(X_train, y_train)}')
print(f'Score on the test set: {lr.score(X_test, y_test)}')

### Logistic Regression Model Using Mean Bucket

X = df_1[['price']]
y = df_1['mean_bucket']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

# Baseline accuracy
y.value_counts(normalize=True)

**Note:** We still have highly unbalanced classes, but it is a bit better than the IRS data.

logreg = LogisticRegression()
logreg.fit(X_train, y_train)

print(f'Score on the training set: {logreg.score(X_train, y_train)}')
print(f'Score on the test set: {logreg.score(X_test, y_test)}')

### Random Forest Model Using Mean Bucket

from sklearn.ensemble import RandomForestClassifier

rf = RandomForestClassifier()
rf.fit(X_train, y_train)

print(f'Score on the training set: {rf.score(X_train, y_train)}')
print(f'Score on the test set: {rf.score(X_test, y_test)}')

### K Nearest Neighbors Using Mean Bucket

from sklearn.neighbors import KNeighborsClassifier

knn = KNeighborsClassifier(n_neighbors=5)
knn.fit(X_train, y_train)

print(f'Score on the training set: {knn.score(X_train, y_train)}')
print(f'Score on the test set: {knn.score(X_test, y_test)}')

## Price Category Models

df_2 = pd.read_csv('../data/combined_4.csv')

df_2.head()

X = df_2[['1', '2', '3', '4', 'total_restaurants']]
y = df_2['mean_household_income']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

### Linear Regression Model

lr = LinearRegression()
lr.fit(X_train, y_train)

print(f'Score on the training set: {lr.score(X_train, y_train)}')
print(f'Score on the test set: {lr.score(X_test, y_test)}')

### Logistic Regression Model

X = df_2[['1', '2', '3', '4']]
y = df_2['median_bucket']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

logreg = LogisticRegression()
logreg.fit(X_train, y_train)

print(f'Score on the training set: {logreg.score(X_train, y_train)}')
print(f'Score on the test set: {logreg.score(X_test, y_test)}')
